Add demo GIF
I want to add an animated GIF that demonstrates Overcommit to the top of
the Overcommit README.md so that people can quickly and easily see an
interesting example of what the project does. To record this, I just set
up a simple repo with a few hooks enabled and committed some basic files
that would fail some of the checks.

I used LICEcap to record and image_optim to optimize.